Implements 2 basic functionalities
1. Create a rank of users in an user network. That is, people that the user don't follow but might find interesting (currently just uses which users follow and/or are mutuals to which other users).
2. A way to return the degrees of separation between two users. That is, two users are considered connected if they are mutuals, so the separation between two users is the length of the shortest path that connects them. It is also convenient to return and verify that path with the most updated online data.

These actions are controlled by a clap-based CLI.

network module defines a graph of user connections and useful functions to interact and update it.

client_utils is a model with useful functions to interact with the client api.

intersection_map is a function that is missing in HashMap but that exists in HashSet. It's based on HashSet::intersection that exists in rust stdlib.